move packages into its own class
